Fixed a bug in memory relocation and added some debug info. The physical memory bitmap takes up to 32 physical pages. Use map_contiguous_virtual_pages.

+++ kernel/boot/init.c
@@ -105,7 +105,7 @@ alloc_idle_TSS (int cpu_num)
   u32 *stk = map_virtual_page (alloc_phys_frame () | 3);
 
 #ifdef USE_VMX
-  pTSS->CR3 = (u32) (get_pdbr () + cpu_num * SANDBOX_KERN_OFFSET);
+  pTSS->CR3 = (u32) (((u32) get_pdbr ()) + cpu_num * SANDBOX_KERN_OFFSET);
 #else
   pTSS->CR3 = (u32) get_pdbr ();
 #endif
